Moved things from GraphSchema to the TypedGraph, made it a class
