+++ src/test/java/com/bio4j/angulillos/Twitter.java
@@ -7,14 +7,14 @@ import java.util.stream.Stream;
 
 public abstract class Twitter<RV,RE>
 extends
-  GraphSchema<Twitter<RV,RE>, RV,RE>
+  TypedGraph<Twitter<RV,RE>, RV,RE>
 {
 
   public Twitter(UntypedGraph<RV,RE> raw) { super(raw); }
 
   public abstract class Vertex<
     V extends Vertex<V>
-  > extends GraphSchema<Twitter<RV,RE>, RV,RE>.Vertex<V> {
+  > extends TypedGraph<Twitter<RV,RE>, RV,RE>.Vertex<V> {
 
     protected Vertex(RV raw, VertexType<V> type) { super(raw, type); }
 
@@ -23,7 +23,10 @@ extends
       E  extends      TypedEdge<V,VertexType<V>, E,ET, ?,?, ?,RV,RE>,
       ET extends TypedEdge.Type<V,VertexType<V>, E,ET, ?,?, ?,RV,RE>
     >
-    Stream<E> outE(ET edgeType) { return graph().outE(self(), edgeType); }
+    Stream<E> outE(ET edgeType) {
+      System.out.println("This is overriden Twitter-graph specific outE");
+      return outE(edgeType);
+    }
   }
